update Feeding model in prisma schema
Removed "amount" field from Feeding model, since the corresponding unit name is a field on the related Food, and changes to that name could make Feeding history nonsensical. Now, a Feeding has a "note" field, which will be prefilled in the creation form with the quantity and unit name of the food used.